requirements.txt
playwright
requests
beautifulsoup4
lxml
python-dotenv
openai
pytest
pytest-mock
